requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
langchain==0.0.430
pymilvus==2.3.0
sentence-transformers==2.2.2
transformers==4.40.0
llama-cpp-python==0.1.52
numpy
python-multipart
requests
tqdm
